chore: move use script test
